refactor: 핸들러 함수 분리 및 forwardRef 적용
