+++ components/radarChart/radarChart.types.ts
@@ -19,8 +19,6 @@ interface IRadarCompoundProps {
 
 interface IRadarChartDraggableProps extends IRadarCompoundProps {
   draggableData: DataPoint[]
-  total: number
-  isRegistered: boolean
 }
 
 interface IRadarChartDefaultProps extends IRadarCompoundProps {
